% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBedFile.R
\name{GetBedFile}
\alias{GetBedFile}
\title{CloudKicker}
\usage{
GetBedFile(genesList, version = "hg19")
}
\arguments{
\item{version}{A caracter string describing the genome version to choose}

\item{geneList}{A vector of gene names in caracter string}
}
\description{
Design to store possible usefull function
}
\details{
This function retrive a dataframe with position of genes transcripts along the genome version you choosed.
}
